WIP: fixed headers (and again)
